Refactor(services): Implement PR feedback.
